split memories out of codex-core
