+++ codex-rs/memories/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "codex-memories"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lib]
+name = "codex_memories"
+path = "src/lib.rs"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+chrono = { workspace = true }
+codex-api = { workspace = true }
+codex-protocol = { workspace = true }
+codex-state = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true, features = ["fs", "macros", "rt"] }
+tracing = { workspace = true }
+uuid = { workspace = true, features = ["serde", "v4", "v5"] }
+
+[dev-dependencies]
+pretty_assertions = { workspace = true }
+tempfile = { workspace = true }
